requirements.txt
absl-py==2.4.0
agents==1.4.0
aiofile==3.9.0
annotated-types==0.7.0
anyio==4.12.1
appnope==0.1.4
asttokens==3.0.1
astunparse==1.6.3
attrs==25.4.0
authlib==1.6.9
beartype==0.22.9
cachetools==7.0.3
caio==0.9.25
certifi==2026.2.25
cffi==2.0.0
charset-normalizer==3.4.5
click==8.3.1
cloudpickle==3.1.2
colorama==0.4.6
comm==0.2.3
cryptography==46.0.5
cyclopts==4.7.0
debugpy==1.8.20
decorator==5.2.1
distro==1.9.0
dnspython==2.8.0
docstring-parser==0.17.0
docutils==0.22.4
email-validator==2.3.0
exceptiongroup==1.3.1
executing==2.2.1
fastmcp==3.1.0
flatbuffers==25.12.19
gast==0.7.0
google-pasta==0.2.0
griffe==1.15.0
grpcio==1.78.0
gym==0.26.2
gym-notices==0.1.0
h11==0.16.0
h5py==3.14.0
httpcore==1.0.9
httpx==0.28.1
httpx-sse==0.4.3
idna==3.11
importlib-metadata==8.7.1
iniconfig==2.3.0
ipykernel==7.2.0
ipython==9.11.0
ipython-pygments-lexers==1.1.1
jaraco-classes==3.4.0
jaraco-context==6.1.0
jaraco-functools==4.4.0
jedi==0.19.2
jiter==0.13.0
jsonref==1.1.0
jsonschema==4.26.0
jsonschema-path==0.4.5
jsonschema-specifications==2025.9.1
jupyter-client==8.8.0
jupyter-core==5.9.1
keras==3.13.2
keyring==25.7.0
libclang==18.1.1
markdown-it-py==4.0.0
matplotlib-inline==0.2.1
mcp==1.26.0
mdurl==0.1.2
ml-dtypes==0.5.4
more-itertools==10.8.0
namex==0.1.0
nest-asyncio==1.6.0
numpy==2.4.2
openai==2.28.0
openai-agents==0.12.1
openapi-pydantic==0.5.1
opentelemetry-api==1.40.0
opt-einsum==3.4.0
optree==0.19.0
packaging==26.0
pandas==3.0.1
parso==0.8.6
pathable==0.5.0
pexpect==4.9.0
platformdirs==4.9.4
pluggy==1.6.0
prompt-toolkit==3.0.52
protobuf==7.34.0
psutil==7.2.2
ptyprocess==0.7.0
pure-eval==0.2.3
py-key-value-aio==0.4.4
pycparser==3.0
pydantic==2.12.5
pydantic-core==2.41.5
pydantic-settings==2.13.1
pygments==2.19.2
pyjwt==2.11.0
pyperclip==1.11.0
pytest==9.0.2
pytest-mock==3.15.1
python-dateutil==2.9.0.post0
python-dotenv==1.2.2
python-multipart==0.0.22
pyyaml==6.0.3
pyzmq==27.1.0
referencing==0.37.0
requests==2.32.5
rich==14.3.3
rich-rst==1.3.2
rpds-py==0.30.0
ruamel-yaml==0.19.1
setuptools==82.0.1
six==1.17.0
sniffio==1.3.1
sse-starlette==3.3.2
stack-data==0.6.3
starlette==0.52.1
tensorflow==2.21.0
termcolor==3.3.0
tornado==6.5.4
tqdm==4.67.3
traitlets==5.14.3
types-requests==2.32.4.20260107
typing-extensions==4.15.0
typing-inspection==0.4.2
uncalled-for==0.2.0
urllib3==2.6.3
uvicorn==0.41.0
watchfiles==1.1.1
wcwidth==0.6.0
websockets==16.0
wheel==0.46.3
wrapt==2.1.2
zipp==3.23.0
